requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.11.1
betterproto==1.2.5
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
configparser==5.3.0
Deprecated==1.2.13
Flask==2.2.2
frozenlist==1.3.3
greenlet==2.0.2
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath==0.82
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.0
opengraph-py3==0.71
Pillow==9.5.0
ping3==4.0.4
polling==0.3.2
priority==2.0.0
pycparser==2.21
pyee==9.0.4
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
python-telegram-bot==20.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
quart==0.18.3
quart-cors==0.6.0
redis==4.4.2
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.28
stringcase==1.2.0
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
webuiapi==0.6.2
wechaty==0.10.7
wechaty-grpc==0.20.19
wechaty-puppet==0.4.23
wechaty-puppet-service==0.8.10
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.12.0

mitmproxy~=9.0.1